# Conflicts:
#	pom.xml

+++ src/main/java/edu/kea/kinoxp/controllers/ScreeningController.java
@@ -42,7 +42,7 @@ public class ScreeningController {
     }
 
     @GetMapping("/screenings/{id}")
-    public String seatSelection(@PathVariable int id, Model model) {
+    public String seatSelection(@PathVariable long id, Model model) {
         model.addAttribute("screening", screeningService.getScreeningByID(id));
         model.addAttribute("seats", screeningService.getOccupiedSeatsByScreeningID(id));
         return "screening/cinema-seats";
